+++ app/helpers/layout_helper.rb
@@ -85,4 +85,7 @@ module LayoutHelper
     end.join(' ').html_safe
   end
 
+  def bootstrap?
+    @css_framework == :bootstrap
+  end
 end