+++ spec/javascripts/app/views/publisher_view_spec.js
@@ -290,7 +290,7 @@ describe("app.views.Publisher", function() {
       it("Show location", function(){
 
         // inserts location to the DOM; it is the location's view element
-        setFixtures('<div id="publisher_textarea_wrapper"></div>');
+        setFixtures('<div id="location_container"></div>');
 
         // creates a fake Locator
         OSM = {};